% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juicr_data.R
\docType{data}
\name{Kam_et_al_2003_Fig2.jpg}
\alias{Kam_et_al_2003_Fig2.jpg}
\title{An example image of a scatterplot figure}
\format{
A raw jpg-formated image
}
\description{
A jpg image of a scatterplot from Figure 2 of Kam, M., Cohen-Gross, S., 
   Khokhlova, I.S., Degen, A.A. and Geffen, E. 2003. Average daily metabolic 
   rate, reproduction and energy allocation during lactation in the Sundevall
   Jird Meriones crassus. Functional Ecology 17:496-503.
}
\note{
\strong{How to use}\cr\cr 
   \code{readImage(system.file("images", "Kam_et_al_2003_Fig2.jpg", package = "juicr"))}
}
\keyword{datasets}
