% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juicr_data.R
\docType{data}
\name{Kortum_and_Acymyan_2013_Fig4.jpg}
\alias{Kortum_and_Acymyan_2013_Fig4.jpg}
\title{An example image of a bar plot figure}
\format{
A raw jpg-formated image
}
\description{
A jpg image of a bar plot from Figure 4 of Kortum, P., and Acymyan, C.Z. 
   2013. How low can you go? Is the System Usability Scale range restricted? 
   Journal of Usability Studies 9:14-24.
}
\note{
\strong{How to use}\cr\cr 
   \code{readImage(system.file("images", "Kortum_and_Acymyan_2013_Fig4.jpg", package = "juicr"))}
}
\keyword{datasets}
